Update dependencies.
Widen version requirement for `rand` to allow update to 0.8.

Remove reference to marsoc crate.

Return to using published crate for bio, now that 0.35 (and 0.36) has
been released.

Widen version range for nix to permit update to 0.20, as required by
pprof 0.4.3.

Widen version ranges for strum and strum_macros to permit latest
version.

Update tests to be forward-compatible with rand 0.8.