Add path = ../... for interdependencies. (#113)
This is in addition to the existing version constraint, which is
required for publishing.  It allows for much easier development when
working on changes which cut across several crates, as well as faster
builds due to not building crates multiple times.

Also add a cargo-deny config to check for accidental introduction of
duplicated dependencies, dependencies with unacceptable licenses, or
dependencies with known security vulnerabilities.