Loosen type constraints on bin_possition etc.
They now permit e.g. searching a &[String] for a &str, rather than
requiring a &String for the search parameter.  The requirement is that
the type in the slice implements Borrow<T> (for which there is a blanket
implementation `borrow(&T) -> &T`).  This is a backwards-compatible
change except in so far as it may alter the compiler's ability to infer
types in some weird cases which are probably bugs anyway.